Reduce tech debt by revamping subscription form
