+++ conf/settings.py
@@ -90,7 +90,6 @@ TEMPLATES = [
                 'django.template.context_processors.debug',
                 'django.template.context_processors.request',
                 'django.template.context_processors.i18n',
-                'core.context_processors.subscribe_form',
                 'core.context_processors.lead_generation_form',
                 'core.context_processors.footer_contact_us_link',
                 'directory_components.context_processors.analytics',
