+++ core/context_processors.py
@@ -1,15 +1,7 @@
-from core.forms import AnonymousSubscribeForm, LeadGenerationForm
+from core.forms import LeadGenerationForm
 from django.utils import translation
 from django.conf import settings
 from directory_constants import urls
-
-
-def subscribe_form(request):
-    return {
-        'subscribe': {
-            'form': AnonymousSubscribeForm(),
-        },
-    }
 
 
 def lead_generation_form(request):
