+++ core/static/core/contact.css
@@ -1,3 +1,3 @@
-@import url("https://fonts.googleapis.com/css?family=Roboto:400,700");@import url("https://fonts.googleapis.com/css?family=Overpass:400,700");html,body,div,span,object,h1,h2,h3,h4,h5,h6,p,blockquote,pre,a,abbr,acronym,address,cite,dfn,em,img,q,small,strong,sub,sup,b,u,i,dl,dt,dd,ol,ul,li,fieldset,form,label,input,legend,table,caption,th,td,article,aside,figure,figcaption,footer,header,nav,section,textarea,video{border:0;box-sizing:border-box;font-size:100%;margin:0;padding:0;vertical-align:baseline}blockquote,q{quotes:none}@media (max-width: 979px){#site-logo{height:auto}}@media (max-width: 860px){.ed-hero-sub-header .ed-hero-sub-header-title{font-size:46px}}@media (max-width: 767px){.ed-hero-sub-header{padding-top:265px}.ed-hero-sub-header .ed-hero-sub-header-title{font-family:"Overpass","Roboto",arial,helvetica,sans-serif;font-size:32px;font-weight:500;line-height:40px;letter-spacing:-0.6px;margin:0 0 0.7em 0;width:540px}#footer-credits{float:left;margin:30px 0 0 0}#intro [class*="span"]{display:block;margin-left:0;width:auto}}@media (max-width: 640px){.browse-companies-container h3,.search-companies-container h3{background-position:right center;padding-left:0;padding-right:51px}.LanguageSelectTracker .language-code:after,.LanguageSelectTracker .language-label{display:none}#companies .browse-companies-container p,#intro .search-companies-container p{margin-bottom:1em;width:100%}#companies .browse-companies-container .button,#intro .search-companies-container .button{position:static}}@media (max-width: 480px){.browse-companies-container p,.search-companies-container p{margin-bottom:25px;width:auto}.browse-companies-container .button,.search-companies-container .button{position:static}.ed-landing-page-form-container{padding:30px 15px}}body{font-family:'Roboto', arial, helvetica, sans-serif}body section{overflow:visible}.ellipsis{overflow:hidden;text-overflow:ellipsis;white-space:nowrap}@media (max-width: 641px){#content .container{margin:0 15px;padding:0 15px}#menu-control{display:none}body.js-enabled #menu-control{display:block}.ed-language-switcher select,.ed-language-switcher p{color:black}body.js-enabled .ed-language-switcher select,.ed-language-switcher p{color:white}.ed-language-switcher{float:left;margin-top:15px}body.js-enabled .ed-language-switcher{float:right;margin-top:5px}}.screen-reader-only{height:1px;left:-10000px;overflow:hidden;position:absolute;top:auto;width:1px}.clearfix:after{visibility:hidden;display:block;font-size:0;content:" ";clear:both;height:0}* html .clearfix{zoom:1}*:first-child+html .clearfix{zoom:1}.a11y{border:0;clip:rect(0 0 0 0);height:1px;margin:-1px;overflow:hidden;padding:0;position:absolute;width:1px}.js-enabled .ed-language-switcher button{display:none}@media (min-width: 1200px){.container,.navbar-static-top .container,.navbar-fixed-top .container,.navbar-fixed-bottom .container{width:1200px;margin:0 auto}}#great-global-header-logo{height:24px}#country-selector-dialog{margin-bottom:0}#country-selector-dialog h2{line-height:1em;margin-bottom:0}#country-selector-dialog select{height:auto}#country-selector-dialog button{color:white}#country-selector-dialog p{margin-bottom:0}#close-country-selector-dialog{padding:0}#content input[type=checkbox].form-control{width:40px}@media (min-width: 641px){main{background:linear-gradient(to right, #fff 70%, #f5f2ed 70%)}#content form input,#content form textarea,#content form select,#content form input[type="checkbox"]+label{width:80%}}#content form button{margin-top:60px;margin-bottom:60px}#id_terms-container .form-label p{margin-bottom:1em}
+@import url("https://fonts.googleapis.com/css?family=Roboto:400,700");@import url("https://fonts.googleapis.com/css?family=Overpass:400,700");html,body,div,span,object,h1,h2,h3,h4,h5,h6,p,blockquote,pre,a,abbr,acronym,address,cite,dfn,em,img,q,small,strong,sub,sup,b,u,i,dl,dt,dd,ol,ul,li,fieldset,form,label,input,legend,table,caption,th,td,article,aside,figure,figcaption,footer,header,nav,section,textarea,video{border:0;box-sizing:border-box;font-size:100%;margin:0;padding:0;vertical-align:baseline}blockquote,q{quotes:none}@media (max-width: 979px){#site-logo{height:auto}}@media (max-width: 860px){.ed-hero-sub-header .ed-hero-sub-header-title{font-size:46px}}@media (max-width: 767px){.ed-hero-sub-header{padding-top:265px}.ed-hero-sub-header .ed-hero-sub-header-title{font-family:"Roboto",arial,helvetica,sans-serif;font-size:32px;font-weight:500;line-height:40px;letter-spacing:-0.6px;margin:0 0 0.7em 0;width:540px}#footer-credits{float:left;margin:30px 0 0 0}#intro [class*="span"]{display:block;margin-left:0;width:auto}}@media (max-width: 640px){.browse-companies-container h3,.search-companies-container h3{background-position:right center;padding-left:0;padding-right:51px}.LanguageSelectTracker .language-code:after,.LanguageSelectTracker .language-label{display:none}#companies .browse-companies-container p,#intro .search-companies-container p{margin-bottom:1em;width:100%}#companies .browse-companies-container .button,#intro .search-companies-container .button{position:static}}@media (max-width: 480px){.browse-companies-container p,.search-companies-container p{margin-bottom:25px;width:auto}.browse-companies-container .button,.search-companies-container .button{position:static}.ed-landing-page-form-container{padding:30px 15px}}body{font-family:'Roboto', arial, helvetica, sans-serif}body section{overflow:visible}.ellipsis{overflow:hidden;text-overflow:ellipsis;white-space:nowrap}@media (max-width: 641px){#content .container{margin:0 15px;padding:0 15px}#menu-control{display:none}body.js-enabled #menu-control{display:block}.ed-language-switcher select,.ed-language-switcher p{color:black}body.js-enabled .ed-language-switcher select,.ed-language-switcher p{color:white}.ed-language-switcher{float:left;margin-top:15px}body.js-enabled .ed-language-switcher{float:right;margin-top:5px}}.screen-reader-only{height:1px;left:-10000px;overflow:hidden;position:absolute;top:auto;width:1px}.clearfix:after{visibility:hidden;display:block;font-size:0;content:" ";clear:both;height:0}* html .clearfix{zoom:1}*:first-child+html .clearfix{zoom:1}.a11y{border:0;clip:rect(0 0 0 0);height:1px;margin:-1px;overflow:hidden;padding:0;position:absolute;width:1px}.js-enabled .ed-language-switcher button{display:none}@media (min-width: 1200px){.container,.navbar-static-top .container,.navbar-fixed-top .container,.navbar-fixed-bottom .container{width:1200px;margin:0 auto}}#great-global-header-logo{height:24px}#country-selector-dialog{margin-bottom:0}#country-selector-dialog h2{line-height:1em;margin-bottom:0}#country-selector-dialog select{height:auto}#country-selector-dialog button{color:white}#country-selector-dialog p{margin-bottom:0}#close-country-selector-dialog{padding:0}#content input[type=checkbox].form-control{width:40px}@media (min-width: 641px){main{background:linear-gradient(to right, #fff 70%, #f5f2ed 70%)}#content form input,#content form textarea,#content form select,#content form input[type="checkbox"]+label{width:80%}}#content form button{margin-top:60px;margin-bottom:60px}#id_terms-container .form-label p{margin-bottom:1em}
 
 /*# sourceMappingURL=maps/contact.css.map */
