--- core/static/core/sass/partials/_variables.scss
@@ -1,10 +1,0 @@
-$asset-path: '../';
-
-$spring-wood: #f5f2ed;
-$mine-shaft: #333333;
-$mid-blue: #00549F;
-$white: #ffffff;
-$focus: #ffbf47;
-
-$font-heading: "Overpass", "Roboto", arial, helvetica, sans-serif;
-$font-family: "Roboto", "Roboto", arial, helvetica, sans-serif;
