+++ core/static/core/sass/subscription.scss
@@ -1,161 +1,16 @@
-@import "partials/variables";
+@import 'partials/base';
+@import 'partials/reset';
+@import "partials/colours";
 
-.subscription {
-  width: 100%;
-  margin-bottom: -30px;
-}
 
 .subscription-form-container {
-  background-color: $spring-wood;
-  padding: 60px;
-  margin: 90px 0 30px;
-  .subscription-register-button-container {
-    margin-bottom: 0;
-    margin-top: 0;
-    text-align: left;
-
-    .button {
-      font-family: $font-heading;
-      width: 230px;
-      margin-top: 19px;
-    }
-  }
-  .subscription-checkbox-container {
-    margin-left: 0;
-  }
-  textarea {
-    min-height: 100px;
-    width: 100%;
-    border: 2px solid $mine-shaft;
-  }
-
-  .title {
-    font-family: $font-heading;
-    font-size: 24px;
-    font-weight: 700;
-    letter-spacing: normal;
-    line-height: 31px;
-    color: $mine-shaft;
-    padding-bottom: 60px;
+  .button {
+    width: 230px;
   }
 }
 
-/* Forms
- *****************/
-
-.subscription-form-inner,
-#company-contact-form {
-  select,
-  textarea,
-  input[type="text"],
-  input[type="password"],
-  input[type="datetime"],
-  input[type="datetime-local"],
-  input[type="date"],
-  input[type="month"],
-  input[type="time"],
-  input[type="week"],
-  input[type="number"],
-  input[type="email"],
-  input[type="url"],
-  input[type="search"],
-  input[type="tel"],
-  input[type="color"],
-  .uneditable-input {
-    /* silly selector to match and overwrite Bootstrap */
-	font-family: $font-family;
-    font-size: 18px;
-    font-weight: 300;
-    line-height: 25px;
-    background-color: $white;
-    border: $mine-shaft solid 1px;
-    border-radius: 0;
-    height: 50px;
-    margin-bottom: 10px;
-    width: 100%;
-
-    &:focus,
-    &:focus:invalid {
-      border-color: $mine-shaft;
-      color: inherit;
-      outline: 3px solid $focus;
-
-      -webkit-box-shadow: none;
-      -moz-box-shadow: none;
-      box-shadow: none;
-
-      -webkit-transition: none;
-      -moz-transition: none;
-      -o-transition: none;
-      transition: none;
-    }
-  }
-
-}
-.checkbox {
-  padding-left: 40px;
-  position: relative;
-  z-index: 0;
-
-  input[type="checkbox"] {
-    height: 22px;
-    left: -1px;
-    margin: 0;
-    opacity: 0;
-    position: absolute;
-    top: 1px;
-    width: 24px;
-    -webkit-appearance: none;
-  }
-
-  input:checked + label:after {
-    content: "";
-    border-color: $mid-blue;
-    border-style: solid;
-    border-width: 0 0 3px 3px;
-    height: 6px;
-    left: 5px;
-    position: absolute;
-    -moz-transform: rotate(-45deg);
-    -o-transform: rotate(-45deg);
-    -webkit-transform: rotate(-45deg);
-    -ms-transform: rotate(-45deg);
-    transform: rotate(-45deg);
-    top: 8px;
-    width: 16px;
-    z-index: -1;
-  }
-
-  input:focus + label:before  {
-    outline: 3px solid $focus;
-  }
-
-  label:before {
-    content: "";
-    border: 1px solid $mine-shaft;
-    display: block;
-    height: 22px;
-    left: 0;
-    position: absolute;
-    top: 1px;
-    width: 24px;
-    z-index: -1;
-  }
-}
-
-.input-field-container {
-  margin-bottom: 15px;
-  label {
-    span {
-      margin-bottom: 5px;
-      display: inline-block;
-    }
-  }
-}
-
-.helptext {
-  font-family: $font-family;
-  font-size: 15px;
-  font-weight: 300;
-  line-height: 23px;
+input, select {
+  width: 100%;
+  padding: 15px;
+  background-color: $white;
 }
