+++ core/tests/test_context_processors.py
@@ -6,22 +6,6 @@ from directory_constants import urls
 from core import context_processors, forms
 
 
-def test_subscribe_form_installed(settings):
-    processors = settings.TEMPLATES[0]['OPTIONS']['context_processors']
-
-    assert 'core.context_processors.subscribe_form' in processors
-
-
-def test_subscribe_form_exposes_form_details(rf):
-    request = rf.get(reverse('index'))
-
-    actual = context_processors.subscribe_form(request)
-
-    assert isinstance(
-        actual['subscribe']['form'], forms.AnonymousSubscribeForm
-    )
-
-
 def test_lead_generation_form_installed(settings):
     processors = settings.TEMPLATES[0]['OPTIONS']['context_processors']
