--- enrolment/templates/anonymous-subscribe-form.html
@@ -1,91 +1,0 @@
-{% load ga360_tracker from directory_components %}
-
-<div class="subscription-form-container">
-    <div class="subscription-form-inner">
-        <div class="grid-row">
-            <h2 class="title column-half-xl">
-                Receive email updates of the latest UK companies in your industry
-            </h2>
-        </div>
-        <div class="row-fluid">
-            {% ga360_tracker "subscribe-form" target="form" %}
-            <form action="{% url 'subscribe' %}" method="post">
-                <div class="grid-row">
-                    <div class="column-full column-third-xl">
-                        <div class="{{ form.full_name.css_classes|default:'input-field-container' }}">
-                            <label for="{{ form.full_name.id_for_label }}">
-                                <span>{{ form.full_name.label }}</span>
-                                {{ form.full_name.errors }}
-                                {{ form.full_name }}
-                            </label>
-                        </div>
-                    </div>
-
-                    <div class="column-full column-third-xl">
-                        <div class="{{ form.email_address.css_classes|default:'input-field-container' }}">
-                            <label for="{{ form.email_address.id_for_label }}">
-                                <span>{{ form.email_address.label }}</span>
-                                {{ form.email_address.errors }}
-                                {{ form.email_address }}
-                            </label>
-                        </div>
-                    </div>
-
-                    <div class="column-full column-third-xl">
-                        <div class="{{ form.sector.css_classes|default:'input-field-container' }}">
-                            <label for="{{ form.sector.id_for_label }}">
-                                <span>{{ form.sector.label }}</span>
-                                {{ form.sector.errors }}
-                                {{ form.sector }}
-                            </label>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="grid-row">
-                    <div class="column-full column-third-xl">
-                        <div class="{{ form.company_name.css_classes|default:'input-field-container' }}">
-                            <label for="{{ form.company_name.id_for_label }}">
-                                <span>{{ form.company_name.label }}</span>
-                                {{ form.company_name.errors }}
-                                {{ form.company_name }}
-                            </label>
-                        </div>
-                    </div>
-
-                    <div class="column-full column-third-xl">
-                        <div class="{{ form.country.css_classes|default:'input-field-container' }}">
-                            <label for="{{ form.country.id_for_label }}">
-                                <span>{{ form.country.label }}</span>
-                                {{ form.country.errors }}
-                                {{ form.country }}
-                            </label>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="row-fluid">
-                    <div
-                        class="subscription-checkbox-container {{ form.terms.css_classes|default:'input-field-container' }}">
-                        {{ form.terms.errors }}
-                        <div class="form-field checkbox">
-                            <input id="{{ form.terms.id_for_label }}" name="{{ form.terms.html_name }}" type="checkbox">
-                            <label for="{{ form.terms.id_for_label }}">
-                                {% ga360_tracker "terms-and-conditions" %}
-                                I agree to the great.gov.uk
-                                    <a target="_self" href="{{ header_footer_urls.terms_and_conditions }}">terms and conditions</a>.
-                                {% endga360_tracker %}
-                            </label>
-                        </div>
-                    </div>
-                    <div class="subscription-register-button-container">
-                        <button class="button button-primary">
-                            Send
-                        </button>
-                    </div>
-                </div>
-            </form>
-            {% endga360_tracker %}
-        </div>
-    </div>
-</div>
