--- enrolment/templates/anonymous-subscribe-success.html
@@ -1,11 +1,0 @@
-{% extends 'enrolment-base.html' %}
-{% load staticfiles %}
-
-{% block content %}
-	<div class="container ed-international-success-container">
-		<h1>Thank you for registering</h1>
-		<p class="ed-international-success-synopsis">
-			We will email you as soon as we publish details of UK companies on the Find a Supplier service.
-		</p>
-	</div>
-{%endblock content %}
