--- enrolment/templates/anonymous-subscribe.html
@@ -1,14 +1,0 @@
-{% extends 'enrolment-base.html' %}
-{% load static from staticfiles %}
-{% block css %}
-    {{ block.super }}
-    <link href="{% static 'company-profile-details.css' %}" media="all" rel="stylesheet" />
-    <link href="{% static 'core/subscription.css' %}" media="all" rel="stylesheet" />
-{% endblock %}
-
-{% block content %}
-    <div class="container">
-        <br>
-        {% include 'anonymous-subscribe-form.html' with subscribe_form=form %}
-    </div>
-{% endblock %}
