+++ find_a_supplier/forms.py
@@ -5,7 +5,11 @@ from directory_validators.common import not_contains_url_or_email
 from directory_forms_api_client.forms import GovNotifyActionMixin
 
 from django.core.validators import EMPTY_VALUES
-from django.forms import HiddenInput, Textarea, TextInput, ValidationError
+from django.forms import (
+    HiddenInput, Select, Textarea, TextInput, ValidationError
+)
+
+from django.utils.translation import ugettext as _
 
 from core.fields import IntegerField
 
@@ -140,3 +144,49 @@ class ContactCompanyForm(GovNotifyActionMixin, forms.Form):
         data = super().serialized_data
         data['sector_label'] = dict(choices.INDUSTRIES)[data['sector']]
         return data
+
+
+class AnonymousSubscribeForm(forms.Form):
+    PLEASE_SELECT_LABEL = _('Please select an industry')
+    TERMS_CONDITIONS_MESSAGE = _(
+        'Tick the box to confirm you agree to the terms and conditions.'
+    )
+    TERMS_CONDITIONS_LABEL = (
+        f'<p>I agree to the <a href="{urls.TERMS_AND_CONDITIONS}" '
+        'target="_blank"> great.gov.uk terms and conditions</a>.</p>'
+    )
+
+    full_name = fields.CharField(label=_('Your name'))
+    email_address = fields.EmailField(label=_('Email address'))
+    sector = fields.ChoiceField(
+        label=_('Industry'),
+        choices=(
+            [['', PLEASE_SELECT_LABEL]] + list(choices.INDUSTRIES)
+        ),
+        widget=Select(attrs={'data-ga-id': 'sector-input'})
+    )
+    company_name = fields.CharField(label=_('Company name'))
+    country = fields.CharField(label=_('Country'))
+    terms = fields.BooleanField(
+        label=TERMS_CONDITIONS_LABEL,
+        error_messages={'required': TERMS_CONDITIONS_MESSAGE}
+    )
+
+
+def serialize_anonymous_subscriber_forms(cleaned_data):
+    """
+    Return the shape directory-api-client expects for saving international
+    buyers.
+
+    @param {dict} cleaned_data - All the fields in `AnonymousSubscribeForm`
+    @returns dict
+
+    """
+
+    return {
+        'name': cleaned_data['full_name'],
+        'email': cleaned_data['email_address'],
+        'sector': cleaned_data['sector'],
+        'company_name': cleaned_data['company_name'],
+        'country': cleaned_data['country'],
+    }
