+++ find_a_supplier/templates/find_a_supplier/search.html
@@ -1,6 +1,6 @@
 {% extends 'core/base-search.html' %}
 
-{% load breadcrumbs from directory_components %}
+{% load ga360_tracker breadcrumbs from directory_components %}
 
 {% block head_title %}Find a UK supplier - GREAT.gov.uk{% endblock %}
 
@@ -20,21 +20,66 @@
 {% endblock %}
 
 {% block subscription %}
-<div class="subscription">
+<div class="subscription background-stone-30 grid-row padding-top-90 padding-bottom-60">
+    <div class="container">
     {% if results %}
-		{% include 'anonymous-subscribe-form.html' with form=subscribe.form %}
-	{% else %}
-		{% if not form.is_bound %}
-		<div id="fassearch-initial-content">
-			<p class="message">Our UK trade profiles will help you contact the right people in the right company.</p>
-			<p class="list-header">Trade profiles give you:</p>
-			<ul>
-				<li>a company summary</li>
-				<li>examples of outstanding projects, goods and services</li>
-				<li>direct contact with a member of a sales or business development team</li>
-			</ul>
-		</div>
-		{% endif %}
-	{% endif %}
+        <div class="subscription-form-container">
+            <div class="subscription-form-inner">
+                <div class="grid-row">
+                    <h2 class="heading-medium margin-bottom-30 margin-top-0">
+                        Receive email updates of the latest UK companies in your industry
+                    </h2>
+                </div>
+                <div class="grid-row">
+                    {% ga360_tracker "subscribe-form" target="form" %}
+                    <form action="{% url 'subscribe' %}" method="post">
+                        <div class="grid-row">
+                            <div class="column-full column-third-xl">
+                                {% include "directory_components/form_widgets/form_field.html" with field=subscribe_form.full_name %}
+                            </div>
+                            <div class="column-full column-third-xl">
+                                {% include "directory_components/form_widgets/form_field.html" with field=subscribe_form.email_address %}
+                            </div>
+                            <div class="column-full column-third-xl">
+                                {% include "directory_components/form_widgets/form_field.html" with field=subscribe_form.sector %}
+                            </div>
+                        </div>
+
+                        <div class="grid-row">
+                            <div class="column-full column-third-xl">
+                                {% include "directory_components/form_widgets/form_field.html" with field=subscribe_form.country %}
+                            </div>
+                            <div class="column-full column-third-xl">
+                                {% include "directory_components/form_widgets/form_field.html" with field=subscribe_form.company_name %}
+                            </div>
+                        </div>
+
+                        <div class="grid-row">
+                            <div clss="column-full">
+                                {% ga360_tracker "terms-and-conditions" target="a" %}
+                                    {% include "directory_components/form_widgets/form_field.html" with field=subscribe_form.terms %}
+                                {% endga360_tracker %}
+                                <button class="button button-primary">Send</button>
+                            </div>
+                        </div>
+                    </form>
+                </div>
+                {% endga360_tracker %}
+            </div>
+        </div>
+    {% else %}
+        {% if not form.is_bound %}
+        <div id="fassearch-initial-content">
+            <p class="message">Our UK trade profiles will help you contact the right people in the right company.</p>
+            <p class="list-header">Trade profiles give you:</p>
+            <ul>
+                <li>a company summary</li>
+                <li>examples of outstanding projects, goods and services</li>
+                <li>direct contact with a member of a sales or business development team</li>
+            </ul>
+        </div>
+        {% endif %}
+    {% endif %} 
+    </div>
 </div>
 {% endblock %}
