+++ find_a_supplier/templates/find_a_supplier/subscribe-success.html
@@ -1,0 +1,35 @@
+{% extends "core/base.html" %}
+{% load static from staticfiles %}
+{% load breadcrumbs message_box ga360_tracker from directory_components %}
+
+{% block head_title %}Find a UK supplier - GREAT.gov.uk{% endblock %}
+
+{% block content %}
+    <section>
+      <div class="container">
+        {% ga360_tracker "breadcrumbs" %}
+			  {% breadcrumbs left_url=services_urls.great_international left_label='Great.gov.uk international' middle_url=services_urls.fas middle_label='Find a Suppluer' right_label='Subscribed' %}
+        {% endga360_tracker %}
+        <div class="grid-row margin-top-30">
+              <div class="column-two-thirds">
+                  <div id="success-message-container">
+                    {% message_box heading="Thank you for registering" description="We will email you as soon as we publish details of UK companies on the Find a Supplier service." heading_level='h1' %}
+                  </div>
+              </div>
+          </div>
+      </div>
+    </section>
+
+    <section id="next-container" class="background-stone-30">
+      <div class="container">
+        <h2 class="heading-large">Next steps</h2>
+        {% ga360_tracker "next-steps" %}
+        <a class="link" href="{% url 'find-a-supplier:search' %}">Browse companies</a>
+        {% endga360_tracker %}
+      </div>
+    </section>
+
+{% endblock %}
+
+
+
