+++ find_a_supplier/templates/find_a_supplier/subscribe.html
@@ -1,0 +1,37 @@
+{% extends 'core/base.html' %}
+{% load static from staticfiles %}
+{% load breadcrumbs from directory_components %}
+{% load ga360_tracker render_form from directory_components %}
+
+
+{% block head_css %}
+    {{ block.super }}
+    <link href="{% static 'core/subscription.css' %}" media="all" rel="stylesheet" />
+{% endblock %}
+
+{% block content %}
+    <div class="container">
+        {% breadcrumbs left_url=services_urls.great_international left_label='Great.gov.uk international' middle_url=services_urls.fas middle_label='Find a Supplier' right_label='Subscribe' %}
+
+
+        <br>
+        <div class="subscription-form-container">
+            <div class="grid-row">
+                <h2 class="heading-large margin-bottom-30 margin-top-0 column-half-xl">
+                    Receive email updates of the latest UK companies in your industry
+                </h2>
+            </div>
+            <div class="grid-row">
+                {% ga360_tracker "subscribe-form" target="form" %}
+                <form action="{% url 'subscribe' %}" method="post" class="column-full column-half-xl">
+                    {% render_form form %}
+                    <button class="button">Send</button>
+                </form>
+            </div>
+            {% endga360_tracker %}
+        </div>
+    </div>
+{% endblock %}
+
+
+
