+++ find_a_supplier/views.py
@@ -113,6 +113,7 @@ class CompanySearchView(
     def get_context_data(self, **kwargs):
         return super().get_context_data(
             show_search_guide=self.should_show_search_guide,
+            subscribe_form=forms.AnonymousSubscribeForm(),
             **kwargs,
         )
 
@@ -290,4 +291,30 @@ class ContactCompanySentView(
             business_unit='FindASupplier',
             site_section='Companies',
             site_subsection='ContactCompanySent',
+        )
+
+
+class SubscribeFormView(CountryDisplayMixin, GA360Mixin, FormView):
+    success_template = 'find_a_supplier/subscribe-success.html'
+    template_name = 'find_a_supplier/subscribe.html'
+    form_class = forms.AnonymousSubscribeForm
+
+    def __init__(self):
+        super().__init__()
+
+        self.set_ga360_payload(
+            page_id='FindASupplierAnonymousSubscribeForm',
+            business_unit='FindASupplier',
+            site_section='AnonymousSubscribe',
+            site_subsection='Form'
+        )
+
+    def form_valid(self, form):
+        data = forms.serialize_anonymous_subscriber_forms(form.cleaned_data)
+        response = api_client.buyer.send_form(data)
+        response.raise_for_status()
+        return TemplateResponse(
+            self.request,
+            self.success_template,
+            self.get_context_data()
         )
