+++ notifications/templates/anonymous_unsubscribe.html
@@ -3,7 +3,7 @@
 {% load ga360_tracker from directory_components %}
 
 {% block css %}
-	<link href="{% static 'company-profile-details.css' %}" media="all" rel="stylesheet" />
+	<link href="{% static 'core/subscription.css' %}" media="all" rel="stylesheet" />
 {% endblock %}
 
 {% block content %}
